增强问题17中的say_number函数，处理更大的数字，并且优化了写法

+++ src/problem17/problem.h
@@ -5,7 +5,7 @@
 class Solution
 {
 	public:
-		std::string say_number(unsigned n);
+		std::string say_number(long n);
 		unsigned count_char(std::string);
 		void answer();
